refactor: split SendFormScreen.kt into smaller focused files (#4004)
Extract large composables from the 1386-line SendFormScreen.kt into
four focused files under ui/screens/send/:

- SendFormComponents.kt: FoldableSection, FadingHorizontalDivider,
  EstimatedNetworkFee, vsClickableBackground (shared helpers)
- SendFormAmountSection.kt: FoldableAmountWidget
- SendFormAddressSection.kt: FoldableDestinationAddressWidget,
  FoldableBondDestinationAddress
- SendFormAssetSection.kt: FoldableAssetWidget

SendFormScreen.kt now contains only sendScreen nav entry,
SendFormScreen, SendFormContent, and the preview. No behavioral
changes; private helpers promoted to internal for cross-file access.
