+++ app/src/main/java/com/vultisig/wallet/ui/screens/send/SendFormScreen.kt
@@ -3,29 +3,14 @@
 package com.vultisig.wallet.ui.screens.send
 
 import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.input.TextFieldState
-import androidx.compose.foundation.text.input.setTextAndPlaceCursorAtEnd
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
@@ -36,46 +21,24 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.focus.FocusManager
 import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.TileMode
 import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import com.vultisig.wallet.R
-import com.vultisig.wallet.ui.components.PasteIcon
-import com.vultisig.wallet.ui.components.PercentageChip
-import com.vultisig.wallet.ui.components.TokenAmountInput
-import com.vultisig.wallet.ui.components.TokenFiatToggle
-import com.vultisig.wallet.ui.components.TokenLogo
 import com.vultisig.wallet.ui.components.UiAlertDialog
-import com.vultisig.wallet.ui.components.UiIcon
 import com.vultisig.wallet.ui.components.UiSpacer
 import com.vultisig.wallet.ui.components.buttons.VsButton
 import com.vultisig.wallet.ui.components.buttons.VsButtonState
-import com.vultisig.wallet.ui.components.inputs.VsTextInputField
-import com.vultisig.wallet.ui.components.inputs.VsTextInputFieldInnerState
-import com.vultisig.wallet.ui.components.library.UiPlaceholderLoader
-import com.vultisig.wallet.ui.components.selectors.ChainSelector
 import com.vultisig.wallet.ui.components.v2.fastselection.contentWithFastSelection
 import com.vultisig.wallet.ui.components.v2.scaffold.V2Scaffold
 import com.vultisig.wallet.ui.models.send.AddressBookType
@@ -85,12 +48,9 @@ import com.vultisig.wallet.ui.models.send.SendFormUiModel
 import com.vultisig.wallet.ui.models.send.SendFormViewModel
 import com.vultisig.wallet.ui.models.send.SendSections
 import com.vultisig.wallet.ui.navigation.Route
-import com.vultisig.wallet.ui.screens.deposit.components.AutoCompoundToggle
-import com.vultisig.wallet.ui.screens.swap.TokenChip
 import com.vultisig.wallet.ui.screens.v2.defi.model.DeFiNavActions
 import com.vultisig.wallet.ui.theme.Theme
 import com.vultisig.wallet.ui.utils.UiText
-import com.vultisig.wallet.ui.utils.VsClipboardService
 import com.vultisig.wallet.ui.utils.asString
 
 internal fun NavGraphBuilder.sendScreen(navController: NavHostController) {
@@ -530,827 +490,6 @@ private fun SendFormContent(
             slippageTexFieldState = slippageFieldState,
             onAutoCompoundCheckedChange = onAutoCompoundCheckedChange,
         )
-    }
-}
-
-@Composable
-private fun FoldableAmountWidget(
-    state: SendFormUiModel,
-    addressFieldState: TextFieldState,
-    amountFocusRequester: FocusRequester = remember { FocusRequester() },
-    onExpandSection: (SendSections) -> Unit,
-    onGasSettingsClick: () -> Unit,
-    tokenAmountFieldState: TextFieldState,
-    fiatAmountFieldState: TextFieldState,
-    focusManager: FocusManager,
-    onSend: () -> Unit,
-    onToogleAmountInputType: (Boolean) -> Unit,
-    onChoosePercentageAmount: (AmountFraction) -> Unit,
-    onChooseMaxTokenAmount: () -> Unit,
-    onAutoCompoundCheckedChange: (Boolean) -> Unit,
-    memoFieldState: TextFieldState,
-    operatorFeeFieldState: TextFieldState,
-    slippageTexFieldState: TextFieldState,
-) {
-    val isCircleMode =
-        state.defiType == DeFiNavActions.DEPOSIT_USDC_CIRCLE ||
-            state.defiType == DeFiNavActions.WITHDRAW_USDC_CIRCLE
-
-    FoldableSection(
-        expanded = isCircleMode || state.expandedSection == SendSections.Amount,
-        onToggle = {
-            if (
-                !isCircleMode && state.isDstAddressComplete && addressFieldState.text.isNotEmpty()
-            ) {
-                onExpandSection(SendSections.Amount)
-            }
-        },
-        expandedTitleActions = {
-            if (!isCircleMode && state.hasGasSettings) {
-                Row(horizontalArrangement = Arrangement.End, modifier = Modifier.weight(1f)) {
-                    UiIcon(
-                        drawableResId = R.drawable.advance_gas_settings,
-                        size = 16.dp,
-                        tint = Theme.v2.colors.text.primary,
-                        onClick = onGasSettingsClick,
-                    )
-                }
-            }
-        },
-        title = stringResource(R.string.send_amount),
-    ) {
-        val contentPadding =
-            Modifier.padding(start = 12.dp, top = 16.dp, end = 12.dp, bottom = 12.dp)
-        Column(
-            modifier =
-                if (isCircleMode) Modifier.fillMaxHeight().then(contentPadding) else contentPadding
-        ) {
-            Box(
-                modifier =
-                    if (isCircleMode) Modifier.weight(1f).fillMaxWidth()
-                    else Modifier.height(211.dp).fillMaxWidth()
-            ) {
-                val primaryAmountText: String
-                val secondaryAmountText: String
-                val primaryFieldState: TextFieldState
-                val secondaryFieldState: TextFieldState
-
-                if (state.usingTokenAmountInput) {
-                    primaryAmountText = state.selectedCoin?.title ?: ""
-                    secondaryAmountText = state.fiatCurrency
-                    primaryFieldState = tokenAmountFieldState
-                    secondaryFieldState = fiatAmountFieldState
-                } else {
-                    primaryAmountText = state.fiatCurrency
-                    secondaryAmountText = state.selectedCoin?.title ?: ""
-                    primaryFieldState = fiatAmountFieldState
-                    secondaryFieldState = tokenAmountFieldState
-                }
-
-                val maxBalance =
-                    if (isCircleMode) state.selectedCoin?.balance?.toBigDecimalOrNull() else null
-
-                val secondaryText =
-                    if (isCircleMode) {
-                        val entered = tokenAmountFieldState.text.toString().toDoubleOrNull() ?: 0.0
-                        val balance = state.selectedCoin?.balance?.toDoubleOrNull() ?: 0.0
-                        val pct =
-                            if (balance > 0.0) (entered / balance * 100).coerceIn(0.0, 100.0)
-                            else 0.0
-                        "%.1f%%".format(pct)
-                    } else {
-                        "${secondaryFieldState.text.ifEmpty { "0" }} $secondaryAmountText"
-                    }
-
-                TokenAmountInput(
-                    primaryFieldState = primaryFieldState,
-                    primaryLabel = primaryAmountText,
-                    secondaryText = secondaryText,
-                    maxBalance = maxBalance,
-                    focusRequester = amountFocusRequester,
-                    onKeyboardAction = {
-                        focusManager.clearFocus()
-                        onSend()
-                    },
-                    modifier =
-                        Modifier.padding(horizontal = 54.dp)
-                            .align(Alignment.Center)
-                            .testTag("SendFormScreen.amountField"),
-                )
-
-                if (!isCircleMode) {
-                    TokenFiatToggle(
-                        isTokenSelected = state.usingTokenAmountInput,
-                        onTokenSelected = { onToogleAmountInputType(it) },
-                        modifier = Modifier.align(Alignment.CenterEnd),
-                    )
-                }
-            }
-
-            UiSpacer(12.dp)
-
-            Row(horizontalArrangement = Arrangement.spacedBy(12.dp)) {
-                state.amountFractionEntries.forEach { fraction ->
-                    PercentageChip(
-                        title = fraction.title.asString(),
-                        isSelected = fraction == state.selectedAmountFraction,
-                        onClick = {
-                            if (fraction == AmountFraction.F100) {
-                                onChooseMaxTokenAmount()
-                            } else {
-                                onChoosePercentageAmount(fraction)
-                            }
-                        },
-                        modifier = Modifier.weight(1f),
-                        isLoading = state.isAmountSelectionLoading,
-                    )
-                }
-            }
-
-            UiSpacer(12.dp)
-
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(6.dp),
-                modifier =
-                    if (isCircleMode) Modifier.padding(all = 16.dp)
-                    else
-                        Modifier.background(
-                                color = Theme.v2.colors.backgrounds.secondary,
-                                shape = RoundedCornerShape(12.dp),
-                            )
-                            .padding(all = 16.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.send_form_balance_available),
-                    style = Theme.brockmann.body.s.medium,
-                    color = Theme.v2.colors.text.primary,
-                )
-
-                val ticker = state.selectedCoin?.title?.let { " $it" } ?: ""
-
-                Text(
-                    text = (state.selectedCoin?.balance ?: "0") + ticker,
-                    style = Theme.brockmann.body.s.medium,
-                    color = Theme.v2.colors.text.secondary,
-                    textAlign = TextAlign.End,
-                    modifier = Modifier.weight(1f),
-                )
-            }
-
-            UiSpacer(12.dp)
-
-            // memo
-            if (state.hasMemo && state.defiType == null) {
-                var isMemoExpanded by remember { mutableStateOf(false) }
-
-                val rotationAngle by
-                    animateFloatAsState(
-                        targetValue = if (isMemoExpanded) 180f else 0f,
-                        animationSpec = tween(durationMillis = 200),
-                        label = "caretRotation",
-                    )
-
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(6.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier =
-                        Modifier.clickable { isMemoExpanded = !isMemoExpanded }
-                            .padding(vertical = 2.dp),
-                ) {
-                    Text(
-                        text = stringResource(R.string.send_form_add_memo),
-                        style = Theme.brockmann.supplementary.caption,
-                        color = Theme.v2.colors.text.tertiary,
-                        modifier = Modifier.weight(1f),
-                    )
-
-                    UiIcon(
-                        drawableResId = R.drawable.ic_caret_down,
-                        tint = Theme.v2.colors.text.primary,
-                        size = 16.dp,
-                        modifier = Modifier.rotate(rotationAngle),
-                    )
-                }
-
-                UiSpacer(12.dp)
-
-                AnimatedVisibility(visible = isMemoExpanded) {
-                    val clipboardData = VsClipboardService.getClipboardData()
-                    VsTextInputField(
-                        textFieldState = memoFieldState,
-                        hint = stringResource(R.string.send_form_enter_memo),
-                        autoCorrectEnabled = false,
-                        trailingIcon = R.drawable.paste,
-                        onTrailingIconClick = {
-                            clipboardData.value
-                                ?.takeIf { it.isNotEmpty() }
-                                ?.let { memoFieldState.setTextAndPlaceCursorAtEnd(text = it) }
-                        },
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-
-                    UiSpacer(12.dp)
-                }
-            }
-
-            if (state.defiType == DeFiNavActions.BOND) {
-                Column(modifier = Modifier.padding(vertical = 2.dp)) {
-                    Text(
-                        text = stringResource(R.string.bond_operator_fees_basis_point),
-                        style = Theme.brockmann.supplementary.caption,
-                        color = Theme.v2.colors.text.tertiary,
-                    )
-
-                    UiSpacer(12.dp)
-
-                    VsTextInputField(
-                        textFieldState = operatorFeeFieldState,
-                        hint = "0",
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-
-                    UiSpacer(12.dp)
-                }
-            }
-
-            if (
-                state.defiType == DeFiNavActions.REDEEM_YRUNE ||
-                    state.defiType == DeFiNavActions.REDEEM_YTCY
-            ) {
-                Column(modifier = Modifier.padding(vertical = 2.dp)) {
-                    Text(
-                        text = stringResource(R.string.deposit_form_operator_slippage_title),
-                        style = Theme.brockmann.supplementary.caption,
-                        color = Theme.v2.colors.text.tertiary,
-                    )
-
-                    UiSpacer(12.dp)
-
-                    VsTextInputField(
-                        textFieldState = slippageTexFieldState,
-                        hint = stringResource(R.string.slippage_hint),
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-
-                    UiSpacer(12.dp)
-                }
-            }
-
-            if (
-                state.defiType == DeFiNavActions.STAKE_TCY ||
-                    state.defiType == DeFiNavActions.STAKE_STCY
-            ) {
-                AutoCompoundToggle(
-                    title = stringResource(R.string.tcy_auto_compound_enable_title),
-                    subtitle = stringResource(R.string.tcy_auto_compound_enable_subtitle),
-                    isChecked = state.isAutocompound,
-                    onCheckedChange = onAutoCompoundCheckedChange,
-                )
-            }
-
-            if (
-                state.defiType == DeFiNavActions.UNSTAKE_TCY ||
-                    state.defiType == DeFiNavActions.UNSTAKE_STCY
-            ) {
-                AutoCompoundToggle(
-                    title = stringResource(R.string.tcy_auto_compound_unstake_title),
-                    subtitle = stringResource(R.string.tcy_auto_compound_unstake_subtitle),
-                    isChecked = state.isAutocompound,
-                    onCheckedChange = onAutoCompoundCheckedChange,
-                )
-            }
-
-            if (state.showGasFee && !isCircleMode) {
-                FadingHorizontalDivider(modifier = Modifier.fillMaxWidth())
-
-                UiSpacer(12.dp)
-
-                EstimatedNetworkFee(
-                    tokenGas = state.totalGas.asString(),
-                    fiatGas = state.estimatedFee.asString(),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun FoldableDestinationAddressWidget(
-    state: SendFormUiModel,
-    onExpandSection: (SendSections) -> Unit,
-    // dst address
-    addressFieldState: TextFieldState,
-    addressFocusRequester: FocusRequester = remember { FocusRequester() },
-    onDstAddressLostFocus: () -> Unit,
-    onSetOutputAddress: (String) -> Unit,
-    onScanDstAddressRequest: () -> Unit,
-    onAddressBookClick: () -> Unit,
-) {
-    FoldableSection(
-        expanded = state.expandedSection == SendSections.Address,
-        complete = state.isDstAddressComplete,
-        title = stringResource(R.string.add_address_address_title),
-        onToggle = { onExpandSection(SendSections.Address) },
-        completeTitleContent = {
-            Text(
-                text = addressFieldState.text.toString(),
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.body.s.medium,
-                maxLines = 1,
-                overflow = TextOverflow.MiddleEllipsis,
-                modifier = Modifier.weight(1f),
-            )
-        },
-    ) {
-        Column(
-            modifier = Modifier.padding(start = 12.dp, top = 16.dp, end = 12.dp, bottom = 12.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.send_from_address),
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.supplementary.caption,
-            )
-
-            UiSpacer(12.dp)
-
-            Column(
-                modifier =
-                    Modifier.fillMaxWidth()
-                        .border(
-                            border =
-                                BorderStroke(width = 1.dp, color = Theme.v2.colors.border.light),
-                            shape = RoundedCornerShape(12.dp),
-                        )
-                        .background(
-                            color = Theme.v2.colors.backgrounds.secondary,
-                            shape = RoundedCornerShape(12.dp),
-                        )
-                        .padding(horizontal = 16.dp, vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(2.dp),
-            ) {
-                Text(
-                    text = state.srcVaultName,
-                    color = Theme.v2.colors.text.primary,
-                    style = Theme.brockmann.supplementary.caption,
-                    maxLines = 1,
-                    overflow = TextOverflow.MiddleEllipsis,
-                )
-
-                Text(
-                    text = state.srcAddress,
-                    color = Theme.v2.colors.text.tertiary,
-                    style = Theme.brockmann.supplementary.caption,
-                    maxLines = 1,
-                    overflow = TextOverflow.MiddleEllipsis,
-                )
-            }
-
-            UiSpacer(16.dp)
-
-            Text(
-                text =
-                    when (state.defiType) {
-                        DeFiNavActions.BOND,
-                        DeFiNavActions.UNBOND -> stringResource(R.string.bond_node_address)
-                        else -> stringResource(R.string.send_to_address)
-                    },
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.supplementary.caption,
-            )
-
-            UiSpacer(12.dp)
-
-            VsTextInputField(
-                textFieldState = addressFieldState,
-                hint = stringResource(R.string.send_to_address_hint),
-                focusRequester = addressFocusRequester,
-                onFocusChanged = {
-                    if (!it) {
-                        onDstAddressLostFocus()
-                    }
-                },
-                keyboardType = KeyboardType.Text,
-                autoCorrectEnabled = false,
-                imeAction = ImeAction.Next,
-                innerState =
-                    if (state.dstAddressError != null) VsTextInputFieldInnerState.Error
-                    else VsTextInputFieldInnerState.Default,
-                footNote = state.dstAddressError?.asString(),
-                modifier = Modifier.fillMaxWidth().testTag("SendFormScreen.addressField"),
-            )
-
-            UiSpacer(16.dp)
-
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                PasteIcon(
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onPaste = onSetOutputAddress,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.camera,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onScanDstAddressRequest,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.ic_bookmark,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onAddressBookClick,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun FoldableBondDestinationAddress(
-    state: SendFormUiModel,
-    onExpandSection: (SendSections) -> Unit,
-    // dst address
-    addressFieldState: TextFieldState,
-    addressFocusRequester: FocusRequester = remember { FocusRequester() },
-    onDstAddressLostFocus: () -> Unit,
-    onSetOutputAddress: (String) -> Unit,
-    onScanDstAddressRequest: () -> Unit,
-    onAddressBookClick: () -> Unit,
-    // bond provider
-    providerFieldState: TextFieldState,
-    onSetOutputProvider: (String) -> Unit,
-    onScanProviderRequest: () -> Unit,
-    onAddressProviderBookClick: () -> Unit,
-) {
-    FoldableSection(
-        expanded = state.expandedSection == SendSections.Address,
-        complete = state.isDstAddressComplete,
-        title = stringResource(R.string.add_address_address_title),
-        onToggle = { onExpandSection(SendSections.Address) },
-        completeTitleContent = {
-            Text(
-                text = addressFieldState.text.toString(),
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.body.s.medium,
-                maxLines = 1,
-                overflow = TextOverflow.MiddleEllipsis,
-                modifier = Modifier.weight(1f),
-            )
-        },
-    ) {
-        Column(
-            modifier = Modifier.padding(start = 12.dp, top = 16.dp, end = 12.dp, bottom = 12.dp)
-        ) {
-            Text(
-                text =
-                    when (state.defiType) {
-                        null,
-                        DeFiNavActions.BOND,
-                        DeFiNavActions.UNBOND -> stringResource(R.string.bond_node_address)
-                        else -> stringResource(R.string.send_to_address)
-                    },
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.supplementary.caption,
-            )
-
-            UiSpacer(12.dp)
-
-            VsTextInputField(
-                textFieldState = addressFieldState,
-                hint = stringResource(R.string.send_to_address_hint),
-                focusRequester = addressFocusRequester,
-                onFocusChanged = {
-                    if (!it) {
-                        onDstAddressLostFocus()
-                    }
-                },
-                keyboardType = KeyboardType.Text,
-                autoCorrectEnabled = false,
-                imeAction = ImeAction.Next,
-                innerState =
-                    if (state.dstAddressError != null) VsTextInputFieldInnerState.Error
-                    else VsTextInputFieldInnerState.Default,
-                footNote = state.dstAddressError?.asString(),
-                modifier = Modifier.fillMaxWidth().testTag("SendFormScreen.addressField"),
-            )
-
-            UiSpacer(16.dp)
-
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                PasteIcon(
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onPaste = onSetOutputAddress,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.camera,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onScanDstAddressRequest,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.ic_bookmark,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onAddressBookClick,
-                )
-            }
-
-            UiSpacer(12.dp)
-
-            Text(
-                text = stringResource(R.string.bond_provider_optional),
-                color = Theme.v2.colors.text.tertiary,
-                style = Theme.brockmann.supplementary.caption,
-            )
-
-            UiSpacer(12.dp)
-
-            VsTextInputField(
-                textFieldState = providerFieldState,
-                hint = stringResource(R.string.send_to_address_hint),
-                keyboardType = KeyboardType.Text,
-                autoCorrectEnabled = false,
-                imeAction = ImeAction.Next,
-                innerState =
-                    if (state.bondProviderError != null) {
-                        VsTextInputFieldInnerState.Error
-                    } else {
-                        VsTextInputFieldInnerState.Default
-                    },
-                footNote = state.bondProviderError?.asString(),
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            UiSpacer(16.dp)
-
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                PasteIcon(
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onPaste = onSetOutputProvider,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.camera,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onScanProviderRequest,
-                )
-
-                UiIcon(
-                    drawableResId = R.drawable.ic_bookmark,
-                    size = 20.dp,
-                    modifier = Modifier.vsClickableBackground().padding(all = 12.dp).weight(1f),
-                    onClick = onAddressProviderBookClick,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun FoldableAssetWidget(
-    state: SendFormUiModel,
-    onExpandSection: (SendSections) -> Unit,
-    onSelectNetworkRequest: () -> Unit,
-    onNetworkDragCancel: () -> Unit,
-    onNetworkDrag: (Offset) -> Unit,
-    onNetworkDragStart: (Offset) -> Unit,
-    onNetworkDragEnd: () -> Unit,
-    onNetworkLongPressStarted: (Offset) -> Unit,
-    onSelectTokenRequest: () -> Unit,
-    onAssetDragCancel: () -> Unit,
-    onAssetDrag: (Offset) -> Unit,
-    onAssetDragStart: (Offset) -> Unit,
-    onAssetDragEnd: () -> Unit,
-    onAssetLongPressStarted: (Offset) -> Unit,
-) {
-    FoldableSection(
-        expanded = state.expandedSection == SendSections.Asset,
-        onToggle = { onExpandSection(SendSections.Asset) },
-        complete = true,
-        title = stringResource(R.string.form_token_selection_asset),
-        completeTitleContent = {
-            Row(modifier = Modifier.weight(1f)) {
-                val selectedToken = state.selectedCoin
-
-                TokenLogo(
-                    errorLogoModifier =
-                        Modifier.size(16.dp).background(Theme.v2.colors.neutrals.n100),
-                    logo = selectedToken?.tokenLogo ?: "",
-                    title = selectedToken?.title ?: "",
-                    modifier = Modifier.size(16.dp),
-                )
-
-                UiSpacer(4.dp)
-
-                Text(
-                    text = selectedToken?.title ?: "",
-                    style = Theme.brockmann.supplementary.caption,
-                    color = Theme.v2.colors.text.tertiary,
-                )
-            }
-        },
-    ) {
-        Column(
-            modifier = Modifier.padding(start = 12.dp, top = 16.dp, end = 12.dp, bottom = 12.dp)
-        ) {
-            val chain = state.selectedCoin?.model?.address?.chain
-            if (chain != null) {
-                Box(modifier = Modifier.testTag("SendFormScreen.chainSelector")) {
-                    ChainSelector(
-                        title = stringResource(R.string.send_from_address),
-                        chain = chain,
-                        onClick = onSelectNetworkRequest,
-                        onDragCancel = onNetworkDragCancel,
-                        onDrag = onNetworkDrag,
-                        onDragStart = onNetworkDragStart,
-                        onDragEnd = onNetworkDragEnd,
-                        onLongPressStarted = onNetworkLongPressStarted,
-                    )
-                }
-            }
-
-            UiSpacer(12.dp)
-
-            Row(modifier = Modifier.fillMaxWidth()) {
-                Box(modifier = Modifier.testTag("SendFormScreen.tokenSelector")) {
-                    TokenChip(
-                        selectedToken = state.selectedCoin,
-                        onSelectTokenClick = onSelectTokenRequest,
-                        onDragCancel = onAssetDragCancel,
-                        onDrag = onAssetDrag,
-                        onDragStart = onAssetDragStart,
-                        onDragEnd = onAssetDragEnd,
-                        onLongPressStarted = onAssetLongPressStarted,
-                    )
-                }
-
-                Column(
-                    horizontalAlignment = Alignment.End,
-                    verticalArrangement = Arrangement.Center,
-                    modifier = Modifier.weight(1f),
-                ) {
-                    state.selectedCoin?.let { token ->
-                        Text(
-                            text =
-                                stringResource(
-                                    R.string.form_token_selection_balance,
-                                    token.balance ?: "",
-                                ),
-                            color = Theme.v2.colors.text.secondary,
-                            style = Theme.brockmann.body.s.medium,
-                            textAlign = TextAlign.End,
-                        )
-
-                        UiSpacer(2.dp)
-
-                        token.fiatValue?.let { fiatValue ->
-                            Text(
-                                text = fiatValue,
-                                textAlign = TextAlign.End,
-                                color = Theme.v2.colors.text.tertiary,
-                                style = Theme.brockmann.supplementary.caption,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-internal fun EstimatedNetworkFee(
-    tokenGas: String,
-    fiatGas: String,
-    isLoading: Boolean = false,
-    title: String = stringResource(R.string.send_form_est_network_fee),
-) {
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        Text(
-            text = title,
-            style = Theme.brockmann.supplementary.footnote,
-            color = Theme.v2.colors.text.tertiary,
-        )
-
-        Column(horizontalAlignment = Alignment.End, modifier = Modifier.weight(1f)) {
-            if (isLoading) {
-                UiPlaceholderLoader(modifier = Modifier.height(20.dp).width(150.dp))
-
-                UiSpacer(6.dp)
-
-                UiPlaceholderLoader(modifier = Modifier.height(20.dp).width(150.dp))
-            } else {
-                Text(
-                    text = tokenGas,
-                    style = Theme.brockmann.body.s.medium,
-                    color = Theme.v2.colors.text.primary,
-                )
-
-                Text(
-                    text = fiatGas,
-                    style = Theme.brockmann.body.s.medium,
-                    color = Theme.v2.colors.text.tertiary,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-internal fun FadingHorizontalDivider(modifier: Modifier = Modifier) {
-    Box(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .height(1.dp)
-                .background(
-                    brush =
-                        Brush.horizontalGradient(
-                            colors =
-                                listOf(
-                                    Theme.v2.colors.backgrounds.secondary.copy(alpha = 0f),
-                                    Color(0xFF284570),
-                                    Theme.v2.colors.backgrounds.secondary.copy(alpha = 0f),
-                                ),
-                            startX = 0f,
-                            endX = Float.POSITIVE_INFINITY,
-                            tileMode = TileMode.Clamp,
-                        )
-                )
-    )
-}
-
-@Composable
-private fun Modifier.vsClickableBackground() =
-    border(
-            border = BorderStroke(width = 1.dp, color = Theme.v2.colors.border.light),
-            shape = RoundedCornerShape(12.dp),
-        )
-        .background(
-            color = Theme.v2.colors.backgrounds.secondary,
-            shape = RoundedCornerShape(12.dp),
-        )
-
-@Composable
-private fun FoldableSection(
-    expanded: Boolean = false,
-    complete: Boolean = false,
-    completeTitleContent: (@Composable RowScope.() -> Unit)? = null,
-    expandedTitleActions: (@Composable RowScope.() -> Unit)? = null,
-    onToggle: () -> Unit = {},
-    title: String,
-    content: @Composable () -> Unit,
-) {
-    Column(
-        modifier =
-            Modifier.border(
-                width = 1.dp,
-                color = Theme.v2.colors.border.normal,
-                shape = RoundedCornerShape(12.dp),
-            )
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(6.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.clickable(onClick = onToggle).padding(all = 16.dp).fillMaxWidth(),
-        ) {
-            Text(
-                text = title,
-                color = Theme.v2.colors.text.primary,
-                style = Theme.brockmann.body.s.medium,
-            )
-
-            if (expanded) {
-                expandedTitleActions?.invoke(this)
-            } else {
-                if (complete) {
-                    completeTitleContent?.invoke(this)
-
-                    UiIcon(
-                        drawableResId = R.drawable.ic_check,
-                        size = 16.dp,
-                        tint = Theme.v2.colors.alerts.success,
-                    )
-
-                    UiSpacer(1.dp)
-
-                    UiIcon(
-                        drawableResId = R.drawable.pencil,
-                        size = 16.dp,
-                        tint = Theme.v2.colors.text.primary,
-                    )
-                }
-            }
-        }
-
-        AnimatedVisibility(visible = expanded) {
-            FadingHorizontalDivider(modifier = Modifier.fillMaxWidth().padding(horizontal = 12.dp))
-
-            content()
-        }
     }
 }
